fixed some css issues. Deleted extra commented ajax codes. sanitize tags.

+++ includes/database/model-table.php
@@ -42,8 +42,6 @@ class FFB_Model_Table {
         add_action( 'wp_ajax_nopriv_addVotesOnRequestList', [$this, 'addVotesOnRequestList'] );
         add_action( 'wp_ajax_removeVotesOnRequestList', [$this, 'removeVotesOnRequestList'] );
         add_action( 'wp_ajax_nopriv_removeVotesOnRequestList', [$this, 'removeVotesOnRequestList'] );
-        // add_action( 'wp_ajax_ffr_deleteTags', [$this, 'ffr_deleteTags'] );
-        // add_action( 'wp_ajax_nopriv_ffr_deleteTags', [$this, 'ffr_deleteTags'] );
         
     }
 
@@ -218,11 +216,12 @@ class FFB_Model_Table {
      */
     public function updateFeatureRequestList() {
         global $wpdb;
+        error_log(print_r($wpdb, 1));
         $table_tag  = $wpdb->prefix . $this->ffr_tags;
         $table_ffr  = $wpdb->prefix . $this->ff_requests_list;
         $id       = (isset($_POST['id']) ? $_POST['id'] : '');
         $parent_id = (isset($_POST['parent_id']) ? $_POST['parent_id'] : '');
-        $title = (isset($_POST['title']) ? $_POST['title'] : '');
+        $title = sanitize_text_field((isset($_POST['title']) ? $_POST['title'] : ''));
         $description = sanitize_textarea_field((isset($_POST['description']) ? $_POST['description'] : ''));
         $status = (isset($_POST['status']) ? $_POST['status'] : '');
         $is_public = (isset($_POST['is_public']) ? $_POST['is_public'] : '');
@@ -230,6 +229,7 @@ class FFB_Model_Table {
         $tagIdDelete = (isset($_POST['tagIdDelete']) ? $_POST['tagIdDelete'] : '');
         $where = ['id' => $id];
 
+        // Update ff_requests_list Table
         $wpdb->update( 
             $table_ffr,
             array( 
@@ -243,11 +243,15 @@ class FFB_Model_Table {
             $where
         );
 
-        foreach($tagIdDelete as $tagid) {
-            $wpdb->delete( $table_tag, array( 'id' => $tagid ) );
+        // Delete Tags
+        if($tagIdDelete) {
+            foreach($tagIdDelete as $tagid) {
+                $wpdb->delete( $table_tag, array( 'id' => $tagid ) );
+            }
         }
 
 
+        // Insert New Tags
         if(!empty($tags)) {
             foreach( $tags as $tag) {
                 $tagid = $tag['id'];
@@ -259,7 +263,7 @@ class FFB_Model_Table {
                     $wpdb->insert(
                         $table_tag,
                         array( 
-                            'name'     => $tag['name'],
+                            'name'     => sanitize_text_field($tag['name']),
                             'slug'     => $tag['slug'],
                             'board_id' => $id,
                         ) 
@@ -270,17 +274,6 @@ class FFB_Model_Table {
         die();
         
     }
-
-    /**
-     * Delete Tags
-     */
-    // public function ffr_deleteTags() {
-    //     global $wpdb;
-    //     $table_tag  = $wpdb->prefix . $this->ffr_tags;
-    //     $id = isset($_POST['tagID']) ? $_POST['tagID'] : '';
-    //     $wpdb->delete( $table_tag, array( 'id' => $id ) );
-    //     die();
-    // }
 
 
     /**
